+++ src/command/ECommand.ts
@@ -1,9 +1,10 @@
 export const enum ECommand {
+    CmdCodeActionAddInclude = 'ahk.nekoHelp.CmdCodeActionAddInclude',
     CmdFindClassRef = 'ahk.nekoHelp.CmdFindClassRef',
     CmdFindComObjConnectRegister = 'ahk.nekoHelp.CmdFindComObjConnectRegister',
-    CmdFindMethodRef = 'ahk.nekoHelp.CmdFindMethodRef',
     CmdFindFuncRef = 'ahk.nekoHelp.CmdFindFuncRef',
     CmdFindLabelRef = 'ahk.nekoHelp.CmdFindLabelRef',
+    CmdFindMethodRef = 'ahk.nekoHelp.CmdFindMethodRef',
     CmdFnAddAhkDoc = 'ahk.nekoHelp.CmdFnAddAhkDoc',
     CmdGotoFuncDef = 'ahk.nekoHelp.CmdGotoFuncDef',
     CompletionMsgBox = 'ahk.nekoHelp.CompletionMsgBox', // TODO