+++ src/extension.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode';
 import { AnalyzeFuncMain } from './command/AnalyzeFunc/AnalyzeThisFunc';
+import { CmdCodeActionAddInclude } from './command/CmdCodeActionAddInclude';
 import { CmdFindClassRef } from './command/CmdFindClassRef';
 import { CmdFindComObjConnectRegister } from './command/CmdFindComObjConnectRegister';
 import { CmdFindFuncRef } from './command/CmdFindFuncRef';
@@ -93,11 +94,12 @@ export function activate(context: vscode.ExtensionContext): void {
         vscode.commands.registerCommand('ahk.nekoHelp.bar', statusBarClick),
         vscode.commands.registerCommand('ahk.nekoHelp.refreshResource', UpdateCacheUi),
         vscode.commands.registerCommand('ahk.nekoHelp.displayLogMessage', displayLogMessageFn), // ctrl+alt+l
+        vscode.commands.registerCommand(ECommand.CmdCodeActionAddInclude, CmdCodeActionAddInclude),
         vscode.commands.registerCommand(ECommand.CmdFindClassRef, CmdFindClassRef),
         vscode.commands.registerCommand(ECommand.CmdFindComObjConnectRegister, CmdFindComObjConnectRegister),
-        vscode.commands.registerCommand(ECommand.CmdFindMethodRef, CmdFindMethodRef),
         vscode.commands.registerCommand(ECommand.CmdFindFuncRef, CmdFindFuncRef),
         vscode.commands.registerCommand(ECommand.CmdFindLabelRef, CmdFindLabelRef),
+        vscode.commands.registerCommand(ECommand.CmdFindMethodRef, CmdFindMethodRef),
         vscode.commands.registerCommand(ECommand.CmdFnAddAhkDoc, CmdFnAddAhkDoc),
         vscode.commands.registerCommand(ECommand.CmdGotoFuncDef, CmdGotoFuncDef),
         vscode.commands.registerCommand(ECommand.ListAllFunc, ListAllFuncMain),