+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [v0.0.48(2023-10)](#v00482023-10)
   - [v0.0.47(2023-07-05)](#v00472023-07-05)
   - [v0.0.46(2023-06-24)](#v00462023-06-24)
   - [v0.0.45(2023-06-22)](#v00452023-06-22)
@@ -27,13 +28,21 @@
   - [0.0.24(2023-02-24)](#00242023-02-24)
   - [\<= 0.0.23(2023-02-17)](#-00232023-02-17)
 
+## v0.0.48(2023-10)
+
+<!-- pnpm 8.9.2-->
+<!-- dprint 0.41.0 -->
+
+- feat: [(#26)](https://github.com/CoffeeChaton/vscode-autohotkey-NekoHelp/issues/26) Go to Definition try support relative path like `#include a.ahk`.
+  ![img](https://user-images.githubusercontent.com/63182963/278785241-1d956407-b223-4360-9cfb-c7402d766f78.gif)
+- fix: [(#27)](https://github.com/CoffeeChaton/vscode-autohotkey-NekoHelp/issues/27) add a CodeAction (yellow light bulb) at `#include` or `#IncludeAgain` range to provider completion.
+  ![img](https://user-images.githubusercontent.com/63182963/278793956-c04f1108-a8e7-4596-b72b-bb603f339640.gif)
+
 ## v0.0.47(2023-07-05)
 
 <!-- pnpm 8.9.0-->
 <!-- dprint 0.37.1 pnpm 8.6.3 -->
 
-- feat: [(#26)](https://github.com/CoffeeChaton/vscode-autohotkey-NekoHelp/issues/26) Go to Definition try support relative path like `#include a.ahk`.
-  ![img](https://user-images.githubusercontent.com/63182963/278785241-1d956407-b223-4360-9cfb-c7402d766f78.gif)
 - fix: parser error of hotkey vs remap exp: `~F12:: foo()` is hotkey call fn, not remap.
 - fix: config error of `"AhkNekoHelp.files.tryParserIncludeOpt": "auto"`
 - fix: syntax-highlight of `and`